chore: move kubernetes function to helm manager
